% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{any2any}
\alias{any2any}
\title{any2any}
\usage{
any2any(testInfo, df1 = NULL, df2 = NULL, N = NULL, n1 = NULL,
  n2 = NULL, esType = NA, var.lor = NA, CIcalc = TRUE, CL = 0.95,
  rID = 0, q = 1, alternative = "two", keepDirection = TRUE,
  keepSign = TRUE, keepSignNames = c("r", "l.r", "u.r", "fisher.z",
  "l.z", "u.z"))
}
\arguments{
\item{df1}{Degrees of freedom}

\item{df2}{NULL or degrees of freedom of the denominator for the f-distribution.}

\item{N}{Number of data points used in calculation of test-statistic.}

\item{n1}{Number of data points in sample 1.}

\item{n2}{Number of data points in sample 2.}

\item{esType}{Type of test statistic. One of: "t", "lm.t", "f", "lm.f", "r", "X2", "Z", "lm.Z"}

\item{CIcalc}{If \code{TRUE} (default) the Confidence Interval for the test statistic in \code{x} will be calculated using the "Confidence limits for noncentral parameters" functions in package (e.g., for type - "t": \link[MBESS]{conf.limits.nct}).}

\item{CL}{Confidence Limit (default: .95).}

\item{rID}{Correlation among predictor values in a linear model.}

\item{q}{Number of predictors in the model.}

\item{alternative}{Alternative hypothesis (default = "two").}

\item{keepSign}{Return effect size with sign of test statistic? (default = TRUE).}

\item{keepSignNames}{Which effect sizes should keep the sign if \code{keepSign = TRUE}? Default is to keep the sign for: "r","l.r","u.r","fisher.z","l.z","u.z".}

\item{st}{Value(s) of a test statistic.}
}
\value{
The effect sizes calculated by \code{compute.es} corresponding to the test statistic(s), with either meta-analytic, or, exact CI.
}
\description{
Converts most common test statistics into most common (signed) effect sizes.
}
\details{
The procedure to calculate a variety of effect sizes is as follows:

\itemize{
\item If \code{CIcalc == FALSE}, \code{package::compute.es} will be used to convert the test statistic to a large number of effect size estimates. The confidence intervals around the effect size estimates will be based meta-analytic estimates of effect size variance (e.g., for type - "t": \link[compute.es]{tes}).
\item If \code{CIcalc == TRUE}, \code{package::MBESS} will be used to calculate the confidence interval for the test statistic based on its noncentral distribution (e.g., for type - "t": \link[MBESS]{conf.limits.nct}). Subsequently the test statistic, as well as its lower and upper confidence limit will each be passed to \code{compute.es} seperately.
\item If \code{keepSign == TRUE} the sign of the test statistic will be copied to all the effect sizes in \code{keepSignNames}.
}
}
\note{
The prefix "lm" is currently disregarded, but will be implemented in future versions to indicate the test statistic is in fact a fixed factor in a linear model.
}
\author{
Fred Hasselman (inspired by RP:P function \code{any2r} by CHJ Hartgerink)
}
