% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{cor_test_fisherZ}
\alias{cor_test_fisherZ}
\title{Fisher Z test for correlations}
\usage{
cor_test_fisherZ(r1 = NULL, r2 = NULL, n1 = NULL, n2 = NULL, p = TRUE,
  Cohens.q = TRUE, conf.level = 0.95, alternative = "two.sided",
  null.value = 0, cor.type = "pearson")
}
\arguments{
\item{r1}{First correlation}

\item{r2}{Second correlation}

\item{n1}{First sample size}

\item{n2}{Second sample size}

\item{p}{Compute p-value? (default = TRUE)}

\item{Cohens.q}{Compute effect size Cohen's q (default = TRUE)}

\item{alternative}{One of "greater", "less", "two.sided" (default)}

\item{alpha}{Alpha evel for significance test}
}
\value{
An `htest.fisherz` object
}
\description{
Fisher Z test for correlations
}
