% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{decide.analysis}
\alias{decide.analysis}
\title{decide.analysis}
\usage{
decide.analysis(ML2.key, studies = NA, tp = NA, doAll = FALSE)
}
\arguments{
\item{ML2.key}{Key}

\item{studies}{studies}

\item{tp}{analysis type}

\item{doAll}{use all studies}
}
\value{
list of studies
}
\description{
decide.analysis
}
