% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRedsRutils.R
\name{decide.EqualVar}
\alias{decide.EqualVar}
\title{decide.Equalvar}
\usage{
decide.EqualVar(vars, labels, key, alpha = 0.05, criterion = 2, group,
  verbose = FALSE)
}
\arguments{
\item{vars}{int}

\item{labels}{int}

\item{key}{int}

\item{alpha}{int}

\item{criterion}{int}

\item{group}{int}

\item{verbose}{int}
}
\description{
decide.Equalvar
}
