% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/fRedsRutils.R
\name{fill_viol}
\alias{fill_viol}
\alias{fill_viol}
\title{fill_viol}
\usage{
fill_viol(gr.df, gr, qtile, probs)

fill_viol(gr.df, gr, qtile, probs)
}
\arguments{
\item{gr.df}{Internal.}

\item{gr}{Internal.}

\item{qtile}{Internal.}

\item{probs}{Internal.}

\item{gr.df}{Internal.}

\item{gr}{Internal.}

\item{qtile}{Internal.}

\item{probs}{Internal.}
}
\value{
A list for \code{\link{vioQtile}}

A list for \code{\link{vioQtile}}
}
\description{
This is adapted from: \url{http://stackoverflow.com/questions/22278951/combining-violin-plot-with-box-plot}

This is adapted from: \url{http://stackoverflow.com/questions/22278951/combining-violin-plot-with-box-plot}
}
\details{
Function to create geom_ploygon calls for \code{\link{vioQtile}}

Function to create geom_ploygon calls for \code{\link{vioQtile}}
}
\seealso{
vioQtile

vioQtile
}
