% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{get.analyses}
\alias{get.analyses}
\title{get.analyses}
\usage{
get.analyses(studies = NA, analysis.type = NA, Nmin.raw = 30,
  Nmin.cond = 15, subset = c("all", "WEIRD", "NON-WEIRD")[1],
  rootdir = normalizePath(paste0(find.package("manylabRs"), "/data")),
  indir = list(RAW.DATA = "RAW.DATA.PRIVATE", MASTERKEY = "MASTERKEY",
  SOURCEINFO = "SOURCEINFO"), outdir = list(ROBJECTS = "ROBJECTS", RESULTS.RDS
  = "RESULTS.RDS"), onlineTables = TRUE)
}
\arguments{
\item{studies}{Numeric vector with unique study IDs listed in the `masteRkey` table (default = all IDs).}

\item{Nmin.raw}{Minimum raw sample size allowed to be included in the analyses.}

\item{Nmin.cond}{Minimum sample size per condition allowed to be included in the analyses.}

\item{tp}{An optional number indicating Global (1), Primary (2, default) or Secondary (3) analyses.}
}
\value{
A list object with analysis results.
}
\description{
get.analyses
}
\details{
Run analyses for (selected) ML2 studies.
}
\examples{
# Get the data from analysis 1 [Huang.1] listed on the \\href{masteRkey spreadsheet}{https://docs.google.com/spreadsheets/d/1fqK3WHwFPMIjNVVvmxpMEjzUETftq_DmP5LzEhXxUHA/}

df <- get.analyses(studies = 1, analysis.type = 1)

# 'raw' pre-filter dataset
head(tbl_df(df$raw.case$Huang.1))

# analysis results
glimpse(tbl_df(df$merged.results$Huang.1))

}
\seealso{
Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.OSFfile}}, \code{\link{get.Order}},
  \code{\link{get.zavCode}}
}
