% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{get.cases}
\alias{get.cases}
\title{get.cases}
\usage{
get.cases(rule, study.vars, study.vars.labels, stat.params)
}
\arguments{
\item{rule}{Internal}

\item{study.vars}{Internal}

\item{study.vars.labels}{Internal}

\item{stat.params}{Internal}
}
\description{
get.cases
}
