% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.chain}
\alias{get.chain}
\alias{get.chain}
\title{get.chain}
\usage{
get.chain(inf)

get.chain(inf)
}
\arguments{
\item{inf}{Internal}

\item{inf}{Internal}
}
\description{
get.chain

get.chain
}
