% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.CSVdata}
\alias{get.CSVdata}
\alias{get.CSVdata}
\title{get.CSVdata}
\usage{
get.CSVdata(path, fID, finishedOnly = TRUE)

get.CSVdata(path, fID, finishedOnly = TRUE)
}
\arguments{
\item{path}{Path to the data.}

\item{finishedOnly}{Only import cases with value of variable \code{Finished = 1} (default).}

\item{files}{A list of \code{.csv} / \code{.xlsx} files containing raw ML2 data.}

\item{path}{Path to the data.}

\item{files}{A list of \code{.csv} / \code{.xlsx} files containing raw ML2 data.}

\item{finishedOnly}{Only import cases with value of variable \code{Finished = 1} (default).}
}
\description{
get.CSVdata

get.CSVdata
}
