% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{get.GoogleSheet}
\alias{get.GoogleSheet}
\title{get.GoogleSheet}
\usage{
get.GoogleSheet(url = NULL, data = c("ML1data", "ML2masteRkey",
  "ML2data")[2], dfCln = FALSE, Sep = ".")
}
\arguments{
\item{url}{Hyperlink to the GoogleSheet, ending in command \code{".../export?format=csv"}.}

\item{data}{If no URL is provided, which dataset? (default = "ML2masteRkey").}

\item{dfCln}{Should the variable names be cleaned (replace spaces and punctuation by a period "."). Default is \code{FALSE}.}

\item{Sep}{Symbol to use when changing column names (default: ".").}
}
\value{
A list object with fields:

\itemize{
\item Returned if \code{dataSet = TRUE}  (default):
\itemize{
\item \code{df}:  A data table generated by \code{\link{tbl_df}} from package \code{dplyr}.
\item \code{info}: Information about the downloaded file including a time stamp, the URL and original row and column names.
}
\item Returned if \code{dataSet = FALSE}:
\itemize{
\item \code{FilePath}: The local path to the downloaded file.
}
}
}
\description{
get.GoogleSheet
}
\seealso{
Other "get." functions: \code{\link{get.OSFfile}},
  \code{\link{get.Order}}, \code{\link{get.analyses}},
  \code{\link{get.zavCode}}
}
\concept{"get." functions}
