% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.info}
\alias{get.info}
\alias{get.info}
\title{get.info}
\usage{
get.info(keytable, cols, subset)

get.info(keytable, cols, subset)
}
\arguments{
\item{keytable}{Internal}

\item{cols}{Internal}

\item{keytable}{Internal}

\item{cols}{Internal}
}
\description{
get.info

get.info
}
