% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/fRedsRutils.R
\name{get.ncpCI}
\alias{get.ncpCI}
\alias{get.ncpCI}
\alias{get.ncpCI}
\title{get.ncpCI}
\usage{
get.ncpCI(x, df1, df2, N, esType, CL = 0.95, keepSign = TRUE,
  keepDirection = TRUE, alternative = "two.sided")

get.ncpCI(x, df1, df2, N, esType, CL = 0.95, keepSign = TRUE,
  keepDirection = TRUE, alternative = "two.sided")

get.ncpCI(x, df1, df2, N, esType, CL = 0.95, keepSign = TRUE,
  keepDirection = TRUE, alternative = "two.sided")
}
\arguments{
\item{x}{A noncentrality parameter.}

\item{df1}{Degrees of freeddom.}

\item{df2}{NULL or degrees of freedom of the denominator for the f-distribution.}

\item{N}{Sample size}

\item{esType}{Type of test statistic. One of: "t", "t.r", lm.t", "f", "lm.f", "r", "X2", "Z", "lm.Z"}

\item{CL}{Confidence Limit (default: .95).}

\item{keepSign}{Return effect size with sign of test statistic? (default = TRUE).}

\item{keepDirection}{Use the information in \code{alternative} to decide on one-sided vs. two-sided confidence intervals. Default is \code{TRUE}. If \code{FALSE}, two-sided CIs will be calclulated irrespective of the direction of the \code{alternative}.}

\item{alternative}{Alternative hypothesis (defult = "two").}

\item{x}{A noncentrality parameter.}

\item{df1}{Degrees of freeddom.}

\item{df2}{NULL or degrees of freedom of the denominator for the f-distribution.}

\item{N}{Sample size}

\item{esType}{Type of test statistic. One of: "t", "t.r", lm.t", "f", "lm.f", "r", "X2", "Z", "lm.Z"}

\item{CL}{Confidence Limit (default: .95).}

\item{keepSign}{Return effect size with sign of test statistic? (default = TRUE).}

\item{keepDirection}{Use the information in \code{alternative} to decide on one-sided vs. two-sided confidence intervals. Default is \code{TRUE}. If \code{FALSE}, two-sided CIs will be calclulated irrespective of the direction of the \code{alternative}.}

\item{alternative}{Alternative hypothesis (defult = "two").}

\item{x}{A noncentrality parameter.}

\item{df1}{Degrees of freeddom.}

\item{df2}{NULL or degrees of freedom of the denominator for the f-distribution.}

\item{N}{Sample size}

\item{esType}{Type of test statistic. One of: "t", "t.r", lm.t", "f", "lm.f", "r", "X2", "Z", "lm.Z"}

\item{CL}{Confidence Limit (default: .95).}

\item{keepSign}{Return effect size with sign of test statistic? (default = TRUE).}

\item{keepDirection}{Use the information in \code{alternative} to decide on one-sided vs. two-sided confidence intervals. Default is \code{TRUE}. If \code{FALSE}, two-sided CIs will be calclulated irrespective of the direction of the \code{alternative}.}

\item{alternative}{Alternative hypothesis (defult = "two").}
}
\description{
get.ncpCI

get.ncpCI

get.ncpCI
}
