% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.Order}
\alias{get.Order}
\alias{get.Order}
\title{get.Order}
\usage{
get.Order(df, S1 = TRUE)

get.Order(df, S1 = TRUE)
}
\arguments{
\item{df}{A ManyLabs2 data frame.}

\item{S1}{Are the data from slate1 (default) or slate2?}

\item{df}{A ManyLabs2 data frame.}

\item{S1}{Are the data from slate1 (default) or slate2?}
}
\value{
A list object with fields:
\itemize{
\item \code{df}: ManyLabs2 data frame in which the Qualtrics study order has been added to each case.
\item \code{Problems}: Cases for which the study order information could not be retrieved.
}

A list object with fields:
\itemize{
\item \code{df}: ManyLabs2 data frame in which the Qualtrics study order has been added to each case.
\item \code{Problems}: Cases for which the study order information could not be retrieved.
}
}
\description{
get.Order

get.Order
}
\seealso{
Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.OSFfile}}, \code{\link{get.analyses}},
  \code{\link{get.zavCode}}

Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.OSFfile}}, \code{\link{get.analyses}},
  \code{\link{get.zavCode}}
}
\author{
Fred Hasselman

Fred Hasselman
}
