% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.OSFfile}
\alias{get.OSFfile}
\alias{get.OSFfile}
\title{get.OSFfile}
\usage{
get.OSFfile(code, dir = tempdir(), scanMethod, downloadMethod = c("httr",
  "downloader", "curl"), dataSet = TRUE, dfCln = FALSE)

get.OSFfile(code, dir = tempdir(), scanMethod, downloadMethod = c("httr",
  "downloader", "curl"), dataSet = TRUE, dfCln = FALSE)
}
\arguments{
\item{code}{Either a full url ("https://osf.io/XXXXX/"), or just the OSF code.}

\item{dir}{Output location (default is \code{tempdir()}).}

\item{scanMethod}{Either \code{readLines} or \code{RCurl}. Leave missing to choose automatically.}

\item{downloadMethod}{One of \code{httr} (default), \code{downloader} or \code{curl}.}

\item{dataSet}{Is the file data set which can be imported using \code{\link{import}} from package \code{rio}?}

\item{dfCln}{Should the variable names be cleaned (replace spaces and punctuation by a period "."). Default is \code{FALSE}.}

\item{code}{Either a full url ("https://osf.io/XXXXX/"), or just the OSF code.}

\item{dir}{Output location (default is \code{tempdir()}).}

\item{scanMethod}{Either \code{readLines} or \code{RCurl}. Leave missing to choose automatically.}

\item{downloadMethod}{One of \code{httr} (default), \code{downloader} or \code{curl}.}

\item{dataSet}{Is the file data set which can be imported using \code{\link{import}} from package \code{rio}?}

\item{dfCln}{Should the variable names be cleaned (replace spaces and punctuation by a period "."). Default is \code{FALSE}.}
}
\value{
A list object with fields:
\itemize{
\item Returned if \code{dataSet = TRUE}  (default):
\itemize{
\item \code{df}:  A data table generated by \code{\link{tbl_df}} from package \code{dplyr}.
\item \code{info}: Information about the downloaded file including a time stamp, the URL and original row and column names.
}
\item Returned if \code{dataSet = FALSE}:
\itemize{
\item \code{FilePath}: The local path to the downloaded file.
}
}

A list object with fields:
\itemize{
\item Returned if \code{dataSet = TRUE}  (default):
\itemize{
\item \code{df}:  A data table generated by \code{\link{tbl_df}} from package \code{dplyr}.
\item \code{info}: Information about the downloaded file including a time stamp, the URL and original row and column names.
}
\item Returned if \code{dataSet = FALSE}:
\itemize{
\item \code{FilePath}: The local path to the downloaded file.
}
}
}
\description{
get.OSFfile

get.OSFfile
}
\details{
Function to download a file hosted on OSF. Modified from code originally written by Sacha Epskamp.

Function to download a file hosted on OSF. Modified from code originally written by Sacha Epskamp.
}
\examples{
#Get the RP:P data hosted on OSF.
\dontrun{dfRPP <- get.OSFfile(code='https://osf.io/fgjvw/', dfCln=TRUE)$df}
#Get the RP:P data hosted on OSF.
\dontrun{dfRPP <- get.OSFfile(code='https://osf.io/fgjvw/', dfCln=TRUE)$df}
}
\seealso{
Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.Order}}, \code{\link{get.analyses}},
  \code{\link{get.zavCode}}

Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.Order}}, \code{\link{get.analyses}},
  \code{\link{get.zavCode}}
}
\author{
Fred Hasselman, based on code by Sasha Epskamp

Fred Hasselman, based on code by Sasha Epskamp
}
