% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{get.output}
\alias{get.output}
\title{get.output}
\usage{
get.output(key, vars, descr, group, analysis, varEqual, test, ESCI,
  test.ConsoleOutput, SourceInfo, stat.test)
}
\arguments{
\item{key}{key}

\item{vars}{vars}

\item{descr}{descr}

\item{group}{group}

\item{analysis}{analysis}

\item{varEqual}{varEqual}

\item{test}{test}

\item{ESCI}{ESCI}

\item{test.ConsoleOutput}{test.ConsoleOutput}

\item{SourceInfo}{SourceInfo}

\item{stat.test}{stat.test}
}
\value{
output
}
\description{
get.output
}
