% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.zavCode}
\alias{get.zavCode}
\alias{get.zavCode}
\title{get.zavCode}
\usage{
get.zavCode(df = NULL, lookup = NULL)

get.zavCode(df = NULL, lookup = NULL)
}
\arguments{
\item{df}{Dataset: ML2_S2}

\item{lookup}{The lookup table}

\item{df}{Dataset: ML2_S2}

\item{lookup}{The lookup table}
}
\value{
The code for each sentence.

The code for each sentence.
}
\description{
get.zavCode

get.zavCode
}
\seealso{
Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.OSFfile}}, \code{\link{get.Order}},
  \code{\link{get.analyses}}

Other "get." functions: \code{\link{get.GoogleSheet}},
  \code{\link{get.OSFfile}}, \code{\link{get.Order}},
  \code{\link{get.analyses}}
}
