% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRedsRutils.R
\name{\%0!0\%}
\alias{\%0!0\%}
\title{Rose tinted infix}
\usage{
x \%0!0\% y
}
\arguments{
\item{x}{If \code{x} is any of \code{Inf,-Inf,NA,NaN,NULL,length(x)==0}, it will return \code{y}; otherwise it returns \code{x}.}

\item{y}{The value to return in case of catastrophy \code{>!<}}
}
\description{
When your functions wear these rose tinted glasses, the world will appear to be a nicer, fluffier place.
}
\examples{
Inf \%0!0\% NA

numeric(0) \%0!0\% ''

NA \%0!0\% 0

NaN \%0!0\% NA

NULL \%0!0\% NA
}
\seealso{
purrr::%||%
}
\author{
Fred Hasselman
}
