% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/inIT.R
\name{in.IT}
\alias{in.IT}
\alias{in.IT}
\title{Initialise It}
\usage{
in.IT(need = NULL, inT = TRUE)

in.IT(need = NULL, inT = TRUE)
}
\arguments{
\item{need}{A vector of package names to be loaded. The wrapper functions have a predefinded \code{need} list and can be used as shortcuts (see details).}

\item{inT}{Logical. If \code{TRUE} (default), packages in \code{need} wil be installed if they are not available on the system.}

\item{need}{A vector of package names to be loaded. The wrapper functions have a predefinded \code{need} list and can be used as shortcuts (see details).}

\item{inT}{Logical. If \code{TRUE} (default), packages in \code{need} wil be installed if they are not available on the system.}
}
\description{
Load and/or install R packages

Load and/or install R packages
}
\details{
\code{in.IT} will check if the Packages in the list argument \code{need} are installed on the system and load them. If \code{inT=TRUE} (default), it will first install the packages if they are not present and then proceed to load them.

\code{in.IT} will check if the Packages in the list argument \code{need} are installed on the system and load them. If \code{inT=TRUE} (default), it will first install the packages if they are not present and then proceed to load them.
}
\examples{
\dontrun{in.IT(c("reshape2", "plyr", "dplyr"))}
in.IT(c("reshape2", "plyr", "dplyr"))
}
\seealso{
Other initialise packages: \code{\link{un.IT}}

Other initialise packages: \code{\link{un.IT}}
}
\author{
Fred Hasselman

Fred Hasselman
}
\keyword{internal}
