% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs.R
\docType{package}
\name{manylabRs}
\alias{manylabRs}
\alias{manylabRs-package}
\title{manylabRs: Analysing data from the Many Labs projects}
\description{
Code and data from the ManyLabs projects.
}
\details{
ML1: \href{https://osf.io/wx7ck/}{Investigating Variation in Replicability}.

ML2: \href{https://osf.io/8cd4r/}{Investigating Variation in Replicability Across Sample and Setting}.
}
\section{`get` functions}{

Functions that get stuff...
}

\section{`varfun` functions}{

The analysis-specific variable functions ...
}

\section{helper functions}{

Functions that help ...
}

