% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{print.cor_test_fisherZ}
\alias{print.cor_test_fisherZ}
\title{print fisherZ}
\usage{
\method{print}{cor_test_fisherZ}(obj)
}
\arguments{
\item{obj}{}
}
\value{
nothing
}
\description{
print fisherZ
}
