% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/fRedsRutils.R
\name{renderHTMLresults}
\alias{renderHTMLresults}
\alias{renderHTMLresults}
\title{renderHTMLresults}
\usage{
renderHTMLresults(pageID)

renderHTMLresults(pageID)
}
\arguments{
\item{pageID}{}

\item{pageID}{}
}
\description{
renderHTMLresults

renderHTMLresults
}
