% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{testScript}
\alias{testScript}
\title{testScript}
\usage{
testScript(studies, tp, saveCSVfile = NA, saveRDSfile = NA,
  subset = c("all", "WEIRD", "NONWEIRD")[1],
  dir.out = "~/Dropbox/Manylabs2/TestOutput")
}
\arguments{
\item{studies}{Unique analysis number(s) from the matsterkey sheet.}

\item{tp}{Analysis type (1 = 'study.global.include', 2 = 'study.primary.include', 3 = 'study.secondary.include').}

\item{saveRDSfile}{Save an RDS file of the output.}
}
\description{
FOR TESTING PURPOSES
}
\keyword{internal}
