% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inIT.R
\name{un.IT}
\alias{un.IT}
\title{Un-initialise It}
\usage{
un.IT(loose, unT = FALSE)
}
\arguments{
\item{loose}{A vector of package names to be unloaded.}

\item{unT}{Logical. If \code{TRUE}, packages in \code{loose} wil be un-installed if they are available on the system.}
}
\description{
Unload and/or uninstall R packages.
}
\details{
\code{un.IT}will check if the Packages in the list argument \code{loose} are installed on the system and unload them. If \code{unT=TRUE} it will first unload the packages if they are loaded, and then proceed to uninstall them.
}
\examples{
\dontrun{un.IT(loose = c("reshape2", "plyr", "dplyr"), unT = FALSE)}
}
\seealso{
Other initialise packages: \code{\link{in.IT}}
}
\author{
Fred Hasselman
}
