% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Alter.4}
\alias{varfun.Alter.4}
\title{varfun.Alter.4}
\usage{
varfun.Alter.4(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#3_alter}
}
\section{Variables}{

  Syllogisms to include for each sample
  INCLUSION PERCENTAGE BASED ON
   FLUENT / DISFLUENT SEPERATELY: 1 5 6
   BOTH: 1 5 6
}

