% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Gati.4}
\alias{varfun.Gati.4}
\alias{varfun.Gati.4}
\alias{varfun.Gati.4}
\title{varfun.Gati.4}
\usage{
varfun.Gati.4(vars)

varfun.Gati.4(vars)

varfun.Gati.4(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#28_gati}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#28_gati}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#28_gati}
}
\note{
This analysis tests moderating effect of presenting Norenzayan first or after.

This analysis tests moderating effect of presenting Norenzayan first or after.

This analysis tests moderating effect of presenting Norenzayan first or after.
}
\references{
Tversky, A., & Gati, I. (1978). Studies of similarity. \strong{Cognition and categorization}, 1, 79-98.

Tversky, A., & Gati, I. (1978). Studies of similarity. \strong{Cognition and categorization}, 1, 79-98.

Tversky, A., & Gati, I. (1978). Studies of similarity. \strong{Cognition and categorization}, 1, 79-98.
}
