% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{varfun.Giessner.1}
\alias{varfun.Giessner.1}
\title{varfun.Giessner.1}
\usage{
varfun.Giessner.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#17_giessner}
}
\section{Variables}{

geis.1.1=long line condition;
geis.2.1=short line condition;
geis.dv_1=dominant;
geis.dv_2=strong;
geis.dv_3=self-confident;
geis.dv_4=control;
geis.dv_5=status;
For all dvs, higher numbers=higher power.
}

