% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Gray.1}
\alias{varfun.Gray.1}
\title{varfun.Gray.1}
\usage{
varfun.Gray.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#22_gray}
}
\section{Variables}{


Adult harms baby scenario; gray1.2=responsibility (adult);  gray1.4=pain (baby).
Baby harms adult scenario; gray2.2=responsibility (baby);  gray2.4=pain (adult)
}

\references{
Gray, K., & Wegner, D. M. (2009). Moral typecasting: divergent perceptions of moral agents and moral patients. Journal of Personality and Social Psychology, 96, 505.
}
