% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Hauser.1}
\alias{varfun.Hauser.1}
\alias{varfun.Hauser.1}
\alias{varfun.Hauser.1}
\title{varfun.Hauser.1}
\usage{
varfun.Hauser.1(vars)

varfun.Hauser.1(vars)

varfun.Hauser.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#11_hauser}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#11_hauser}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#11_hauser}
}
\section{Variables}{

  haus1.1t = timing (side effect scenario);
  haus2.1t = timing (greater good scenario);
  haus1.2=previous experience (drop if 1 (yes));
  haus2.2=previous experience (drop if 1 (yes));
  haus1.1=morally permissible (side effect scenario; Yes=1);
  haus2.1=morally permissible (greater good scenario; Yes=1).


  haus1.1t = timing (side effect scenario);
  haus2.1t = timing (greater good scenario);
  haus1.2=previous experience (drop if 1 (yes));
  haus2.2=previous experience (drop if 1 (yes));
  haus1.1=morally permissible (side effect scenario; Yes=1);
  haus2.1=morally permissible (greater good scenario; Yes=1).


  haus1.1t = timing (side effect scenario);
  haus2.1t = timing (greater good scenario);
  haus1.2=previous experience (drop if 1 (yes));
  haus2.2=previous experience (drop if 1 (yes));
  haus1.1=morally permissible (side effect scenario; Yes=1);
  haus2.1=morally permissible (greater good scenario; Yes=1).
}

