% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Hauser.2}
\alias{varfun.Hauser.2}
\alias{varfun.Hauser.2}
\alias{varfun.Hauser.2}
\title{varfun.Hauser.2}
\usage{
varfun.Hauser.2(vars)

varfun.Hauser.2(vars)

varfun.Hauser.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#17_hauser}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#17_hauser}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#17_hauser}
}
\section{Variables}{

hauser3.1=morality judgment (greater good condition);
hauser4.1=morality judgment (foreseen side-effect condition; for both, yes=1, no=2.)
haus3.2 and haus4.2=previous experience (yes=1, no=2);
haus3.1t_3=timing (greater good);
haus4.1t_3=timing (side effect).


hauser3.1=morality judgment (greater good condition);
hauser4.1=morality judgment (foreseen side-effect condition; for both, yes=1, no=2.)
haus3.2 and haus4.2=previous experience (yes=1, no=2);
haus3.1t_3=timing (greater good);
haus4.1t_3=timing (side effect).


hauser3.1=morality judgment (greater good condition);
hauser4.1=morality judgment (foreseen side-effect condition; for both, yes=1, no=2.)
haus3.2 and haus4.2=previous experience (yes=1, no=2);
haus3.1t_3=timing (greater good);
haus4.1t_3=timing (side effect).
}

\references{
Hauser, M., Cushman, F., Young, L., Kang-Xing Jin, R., & Mikhail, J. (2007). A dissociation between moral judgments and justifications. Mind & Language, 22, 1-21.

Hauser, M., Cushman, F., Young, L., Kang-Xing Jin, R., & Mikhail, J. (2007). A dissociation between moral judgments and justifications. Mind & Language, 22, 1-21.

Hauser, M., Cushman, F., Young, L., Kang-Xing Jin, R., & Mikhail, J. (2007). A dissociation between moral judgments and justifications. Mind & Language, 22, 1-21.
}
