% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/ML2_variable_functions.R
\name{varfun.Hsee.1}
\alias{varfun.Hsee.1}
\alias{varfun.Hsee.1}
\title{varfun.Hsee.1}
\usage{
varfun.Hsee.1(vars)

varfun.Hsee.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#21_hsee}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#21_hsee}
}
\section{Variables}{

hsee1.1=generosity ($90 scarf condition);
hsee2.1=generosity ($110 coat condition);
for both, higher numbers=higher generosity


hsee1.1=generosity ($90 scarf condition);
hsee2.1=generosity ($110 coat condition);
for both, higher numbers=higher generosity
}

\references{
Hsee, C. K. (1998). Less is better: When low-value options are valued more highly than high-value options. Journal of Behavioral Decision Making, 11, 107-121.

Hsee, C. K. (1998). Less is better: When low-value options are valued more highly than high-value options. Journal of Behavioral Decision Making, 11, 107-121.
}
