% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Huang.1}
\alias{varfun.Huang.1}
\alias{varfun.Huang.1}
\alias{varfun.Huang.1}
\title{Huang.1}
\usage{
varfun.Huang.1(vars)

varfun.Huang.1(vars)

varfun.Huang.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis A \code{list} object containing fields \strong{High}, \strong{Low} and \strong{N}

Dataset ready for analysis A \code{list} object containing fields \strong{High}, \strong{Low} and \strong{N}

Dataset ready for analysis A \code{list} object containing fields \strong{High}, \strong{Low} and \strong{N}
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#1_huang}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#1_huang}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#1_huang}
}
\section{Variables}{


huan1.1_Y1: Y position of the mouse (High SES condition).
huan2.1_Y1: Y position of the mouse (Low SES).
huan1.1_R0 and huan2.1_R0 indicate for each condition whether a click was inside the map (1) or outside (0).

For each condition a participant must have clicked inside the map (=1) to be included in the analysis.



huan1.1_Y1: Y position of the mouse (High SES condition).
huan2.1_Y1: Y position of the mouse (Low SES).
huan1.1_R0 and huan2.1_R0 indicate for each condition whether a click was inside the map (1) or outside (0).

For each condition a participant must have clicked inside the map (=1) to be included in the analysis.



huan1.1_Y1: Y position of the mouse (High SES condition).
huan2.1_Y1: Y position of the mouse (Low SES).
huan1.1_R0 and huan2.1_R0 indicate for each condition whether a click was inside the map (1) or outside (0).

For each condition a participant must have clicked inside the map (=1) to be included in the analysis.
}

