% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/ML2_variable_functions.R
\name{varfun.Inbar.1}
\alias{varfun.Inbar.1}
\alias{varfun.Inbar.1}
\title{varfun.Inbar.1}
\usage{
varfun.Inbar.1(vars)

varfun.Inbar.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#8_inbar}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#8_inbar}
}
\section{Variables}{

disg1.11,disg1.12,disg2.10,disg2.12,disg2.13;

responses on the DS-R are scored as follows: True 1, False 0; Not disgusting 0, Slightly disgusting 0.5, Very disgusting 1


disg1.11,disg1.12,disg2.10,disg2.12,disg2.13;

responses on the DS-R are scored as follows: True 1, False 0; Not disgusting 0, Slightly disgusting 0.5, Very disgusting 1
}

