% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/ML2_variable_functions.R
\name{varfun.Knobe.1}
\alias{varfun.Knobe.1}
\alias{varfun.Knobe.1}
\title{varfun.Knobe.1}
\usage{
varfun.Knobe.1(vars)

varfun.Knobe.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#27_knobe}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#27_knobe}
}
\section{Variables}{

  knob1.3=intentionality (help condition);
  knob2.3=intentionality (harm condition);
  for both, higher numbers=higher intentionality


  knob1.3=intentionality (help condition);
  knob2.3=intentionality (harm condition);
  for both, higher numbers=higher intentionality
}

\references{
Knobe, J. (2003). Intentional action and side effects in ordinary language. Analysis, 63, 190-193.

Knobe, J. (2003). Intentional action and side effects in ordinary language. Analysis, 63, 190-193.
}
