% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{varfun.Knobe.2}
\alias{varfun.Knobe.2}
\title{varfun.Knobe.2}
\usage{
varfun.Knobe.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#27_knobe}
}
\section{Variables}{

  knob1.4=intentionality (praise condition);
  knob2.4=intentionality (blame condition);
  for both, higher numbers=higher intentionality
}

\references{
Knobe, J. (2003). Intentional action and side effects in ordinary language. Analysis, 63, 190-193.
}
