% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{varfun.Norenzayan.2}
\alias{varfun.Norenzayan.2}
\title{varfun.Norenzayan.2}
\usage{
varfun.Norenzayan.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#20_norenzayan}
}
\note{
This analysis tests moderating effect of presenting Gati first or after.
}
\references{
Norenzayan, A., Smith, E. E., Kim, B. J., & Nisbett, R. E. (2002). Cultural preferences for formal versus intuitive reasoning. Cognitive Science, 26, 653-684.
}
