% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Risen.1}
\alias{varfun.Risen.1}
\alias{varfun.Risen.1}
\alias{varfun.Risen.1}
\title{varfun.Risen.1}
\usage{
varfun.Risen.1(vars)

varfun.Risen.1(vars)

varfun.Risen.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#18_risen}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#18_risen}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#18_risen}
}
\section{Variables}{

rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);
for both, higher numbers=higher likelihood
Variable = "ex.subjp" which asked if participants were recruited through a university subject pool. 1 = yes, 2 = no.


rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);
for both, higher numbers=higher likelihood
Variable = "ex.subjp" which asked if participants were recruited through a university subject pool. 1 = yes, 2 = no.


rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);
for both, higher numbers=higher likelihood
Variable = "ex.subjp" which asked if participants were recruited through a university subject pool. 1 = yes, 2 = no.
}

\references{
Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. \strong{Journal of Personality and Social Psychology}, 95, 293.

Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. \strong{Journal of Personality and Social Psychology}, 95, 293.

Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. \strong{Journal of Personality and Social Psychology}, 95, 293.
}
