% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Risen.2}
\alias{varfun.Risen.2}
\alias{varfun.Risen.2}
\alias{varfun.Risen.2}
\title{varfun.Risen.2}
\usage{
varfun.Risen.2(vars)

varfun.Risen.2(vars)

varfun.Risen.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#18_risen}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#18_risen}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#18_risen}
}
\section{Variables}{

rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);

for both, higher numbers=higher likelihood   All participants that answer the dependent measure will be included in analysis.
The primary confirmatory test for comparing the original and replication effect size will be based on only the samples using undergraduate students.


rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);

for both, higher numbers=higher likelihood   All participants that answer the dependent measure will be included in analysis.
The primary confirmatory test for comparing the original and replication effect size will be based on only the samples using undergraduate students.


rise1.3=likelihood that the professor will call on you (unprepared condition);
rise2.3=likelihood that the professor will call on you (prepared condition);

for both, higher numbers=higher likelihood   All participants that answer the dependent measure will be included in analysis.
The primary confirmatory test for comparing the original and replication effect size will be based on only the samples using undergraduate students.
}

\references{
Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. Journal of Personality and Social Psychology, 95, 293.

Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. Journal of Personality and Social Psychology, 95, 293.

Risen, J. L., & Gilovich, T. (2008). Why people are reluctant to tempt fate. Journal of Personality and Social Psychology, 95, 293.
}
