% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Rottenstreich.1}
\alias{varfun.Rottenstreich.1}
\alias{varfun.Rottenstreich.1}
\alias{varfun.Rottenstreich.1}
\title{varfun.Rottenstreich.1}
\usage{
varfun.Rottenstreich.1(vars)

varfun.Rottenstreich.1(vars)

varfun.Rottenstreich.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.

 @export}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.

 @export}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.

 @export}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#5_rottenstreich}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#5_rottenstreich}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#5_rottenstreich}
}
