% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Savani.1}
\alias{varfun.Savani.1}
\alias{varfun.Savani.1}
\alias{varfun.Savani.1}
\title{varfun.Savani.1}
\usage{
varfun.Savani.1(vars)

varfun.Savani.1(vars)

varfun.Savani.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#19_savani}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#19_savani}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#19_savani}
}
\section{Variables}{

sava1.N=interpersonal actions;

sava2.N=personal actions;

'sava1.4', 'sava1.5', 'sava1.9', 'sava1.10', 'sava1.15', 'sava1.16' , 'sava1.21', 'sava1.22', 'sava1.27', 'sava1.28', 'sava1.33', 'sava1.34',  'sava1.38', 'sava1.39', 'sava1.43', 'sava1.44'

sava1.4=choice (buy a gift; 1=choice; 2=no choice);

sava1.5=importance (buy a gift);

sava1.9=choice (take a friend at the restaurant; 1=choice; 2=no choice);

sava1.10=importance (restaurant);

sava1.15=choice (trip; 1=choice, 2 and 3 = no choice);

sava1.16=importance (trip);

sava1.21=choice (dinner; 1=choice, 2 and 3 = no choice);

sava1.22=importance (dinner);

sava1.27=choice (errand; 1=choice, 2 and 3 = no choice);

sava1.28=importance (errand);

sava1.33=choice (help, 1=choice, 2 & 3 = no choice);

sava1.34=importance (help);

sava1.38=choice (advice, 1=choice, 2 & 3 = no choice);

sava1.39=importance (advice);

sava1.43=choice (friends, 1=choice, 2 & 3 = no choice);

sava1.44=importance (friends);


sava2.4=choice (buy for yourself; 1=choice; 2=no choice);

sava2.5=importance (buy for yourself);

sava2.9=choice (at the restaurant by yourself; 1=choice; 2=no choice);

sava2.10=importance (restaurant by yourself);

sava2.15=choice (trip alone; 1=choice, 2 and 3 = no choice);

sava2.16=importance (trip alone);

sava2.21=choice (out for dinner; 1=choice, 2 and 3 = no choice);

sava2.22=importance (out for dinner);

sava2.27=choice (errand for yourself; 1=choice, 2 and 3 = no choice);

sava2.28=importance (errand for yourself);

sava2.33=choice (ask for help, 1=choice, 2 & 3 = no choice);

sava2.34=importance (ask for help);

sava2.38=choice (take a course, 1=choice, 2 & 3 = no choice);

sava2.39=importance (take a course);

sava2.43=choice (friends, 1=choice, 2 & 3 = no choice);

sava2.44=importance (friends);


For all importance items: higher numbers=higher importance

we will only include university data collections in the primary confirmatory analysis to be compared with the original effect sizes.

Data for all participants will be included to examine variability across sample and setting.
However, participants must respond to all choice and importance of choice questions to be included in the analysis.


sava1.N=interpersonal actions;

sava2.N=personal actions;

'sava1.4', 'sava1.5', 'sava1.9', 'sava1.10', 'sava1.15', 'sava1.16' , 'sava1.21', 'sava1.22', 'sava1.27', 'sava1.28', 'sava1.33', 'sava1.34',  'sava1.38', 'sava1.39', 'sava1.43', 'sava1.44'

sava1.4=choice (buy a gift; 1=choice; 2=no choice);

sava1.5=importance (buy a gift);

sava1.9=choice (take a friend at the restaurant; 1=choice; 2=no choice);

sava1.10=importance (restaurant);

sava1.15=choice (trip; 1=choice, 2 and 3 = no choice);

sava1.16=importance (trip);

sava1.21=choice (dinner; 1=choice, 2 and 3 = no choice);

sava1.22=importance (dinner);

sava1.27=choice (errand; 1=choice, 2 and 3 = no choice);

sava1.28=importance (errand);

sava1.33=choice (help, 1=choice, 2 & 3 = no choice);

sava1.34=importance (help);

sava1.38=choice (advice, 1=choice, 2 & 3 = no choice);

sava1.39=importance (advice);

sava1.43=choice (friends, 1=choice, 2 & 3 = no choice);

sava1.44=importance (friends);


sava2.4=choice (buy for yourself; 1=choice; 2=no choice);

sava2.5=importance (buy for yourself);

sava2.9=choice (at the restaurant by yourself; 1=choice; 2=no choice);

sava2.10=importance (restaurant by yourself);

sava2.15=choice (trip alone; 1=choice, 2 and 3 = no choice);

sava2.16=importance (trip alone);

sava2.21=choice (out for dinner; 1=choice, 2 and 3 = no choice);

sava2.22=importance (out for dinner);

sava2.27=choice (errand for yourself; 1=choice, 2 and 3 = no choice);

sava2.28=importance (errand for yourself);

sava2.33=choice (ask for help, 1=choice, 2 & 3 = no choice);

sava2.34=importance (ask for help);

sava2.38=choice (take a course, 1=choice, 2 & 3 = no choice);

sava2.39=importance (take a course);

sava2.43=choice (friends, 1=choice, 2 & 3 = no choice);

sava2.44=importance (friends);


For all importance items: higher numbers=higher importance

we will only include university data collections in the primary confirmatory analysis to be compared with the original effect sizes.

Data for all participants will be included to examine variability across sample and setting.
However, participants must respond to all choice and importance of choice questions to be included in the analysis.


sava1.N=interpersonal actions;

sava2.N=personal actions;

'sava1.4', 'sava1.5', 'sava1.9', 'sava1.10', 'sava1.15', 'sava1.16' , 'sava1.21', 'sava1.22', 'sava1.27', 'sava1.28', 'sava1.33', 'sava1.34',  'sava1.38', 'sava1.39', 'sava1.43', 'sava1.44'

sava1.4=choice (buy a gift; 1=choice; 2=no choice);

sava1.5=importance (buy a gift);

sava1.9=choice (take a friend at the restaurant; 1=choice; 2=no choice);

sava1.10=importance (restaurant);

sava1.15=choice (trip; 1=choice, 2 and 3 = no choice);

sava1.16=importance (trip);

sava1.21=choice (dinner; 1=choice, 2 and 3 = no choice);

sava1.22=importance (dinner);

sava1.27=choice (errand; 1=choice, 2 and 3 = no choice);

sava1.28=importance (errand);

sava1.33=choice (help, 1=choice, 2 & 3 = no choice);

sava1.34=importance (help);

sava1.38=choice (advice, 1=choice, 2 & 3 = no choice);

sava1.39=importance (advice);

sava1.43=choice (friends, 1=choice, 2 & 3 = no choice);

sava1.44=importance (friends);


sava2.4=choice (buy for yourself; 1=choice; 2=no choice);

sava2.5=importance (buy for yourself);

sava2.9=choice (at the restaurant by yourself; 1=choice; 2=no choice);

sava2.10=importance (restaurant by yourself);

sava2.15=choice (trip alone; 1=choice, 2 and 3 = no choice);

sava2.16=importance (trip alone);

sava2.21=choice (out for dinner; 1=choice, 2 and 3 = no choice);

sava2.22=importance (out for dinner);

sava2.27=choice (errand for yourself; 1=choice, 2 and 3 = no choice);

sava2.28=importance (errand for yourself);

sava2.33=choice (ask for help, 1=choice, 2 & 3 = no choice);

sava2.34=importance (ask for help);

sava2.38=choice (take a course, 1=choice, 2 & 3 = no choice);

sava2.39=importance (take a course);

sava2.43=choice (friends, 1=choice, 2 & 3 = no choice);

sava2.44=importance (friends);


For all importance items: higher numbers=higher importance

we will only include university data collections in the primary confirmatory analysis to be compared with the original effect sizes.

Data for all participants will be included to examine variability across sample and setting.
However, participants must respond to all choice and importance of choice questions to be included in the analysis.
}

