% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{varfun.Schwarz.1}
\alias{varfun.Schwarz.1}
\title{varfun.Schwarz.1}
\usage{
varfun.Schwarz.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#24_schwarz}
}
\section{Variables}{

schw1.1=life sat (first);
schw1.2=partner satisfaction (second);
schw2.1=partner satisfaction (first);
schw2.2=life sat (second).
for all, higher numbers=higher satisfaction
}

\references{
Schwarz, N., Strack, F., & Mai, H. P. (1991). Assimilation and contrast effects in part-whole question sequences: A conversational logic analysis. \strong{Public Opinion Quarterly, 55}, 3-23.
}
