% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/ML2_variable_functions.R
\name{varfun.Schwarz.2}
\alias{varfun.Schwarz.2}
\alias{varfun.Schwarz.2}
\title{varfun.Schwarz.2}
\usage{
varfun.Schwarz.2(vars)

varfun.Schwarz.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#24_schwarz}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#24_schwarz}
}
\details{
\strong{Analysis plan:} We will compute the correlation between responses to the general and
 specific question in each item order condition, and then compare the correlations using
 the Fisher r-to-z transformation. Participants with valid responses to both items will
 be included in the analysis.


 @export

\strong{Analysis plan:} We will compute the correlation between responses to the general and
 specific question in each item order condition, and then compare the correlations using
 the Fisher r-to-z transformation. Participants with valid responses to both items will
 be included in the analysis.


 @export
}
\references{
Schwarz, N., Strack, F., & Mai, H. P. (1991). Assimilation and contrast effects in part-whole question sequences: A conversational logic analysis. \strong{Public Opinion Quarterly, 55}, 3-23.

Schwarz, N., Strack, F., & Mai, H. P. (1991). Assimilation and contrast effects in part-whole question sequences: A conversational logic analysis. \strong{Public Opinion Quarterly, 55}, 3-23.
}
