% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Shafir.1}
\alias{varfun.Shafir.1}
\alias{varfun.Shafir.1}
\alias{varfun.Shafir.1}
\title{varfun.Shafir.1}
\usage{
varfun.Shafir.1(vars)

varfun.Shafir.1(vars)

varfun.Shafir.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#25_shafir}

\strong{CHANGED} The analysis in the orignal article was likely as follows:

- Count number of Parent B choices in both conditions
- Sum the proportions
- Divide by 2 and test against proportion = .5

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#25_shafir}

\strong{CHANGED} The analysis in the orignal article was likely as follows:

- Count number of Parent B choices in both conditions
- Sum the proportions
- Divide by 2 and test against proportion = .5

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#25_shafir}

\strong{CHANGED} The analysis in the orignal article was likely as follows:

- Count number of Parent B choices in both conditions
- Sum the proportions
- Divide by 2 and test against proportion = .5
}
\section{Variables}{

shaf1.1=choice (award condition; Parent A=1, Parent B=2);
shaf2.1=choice (deny condition; Parent A=1, Parent B=2)


shaf1.1=choice (award condition; Parent A=1, Parent B=2);
shaf2.1=choice (deny condition; Parent A=1, Parent B=2)


shaf1.1=choice (award condition; Parent A=1, Parent B=2);
shaf2.1=choice (deny condition; Parent A=1, Parent B=2)
}

\references{
Shafir, E. (1993). Choosing versus rejecting: Why some options are both better and worse than others. Memory & Cognition, 21, 546-556.

Shafir, E. (1993). Choosing versus rejecting: Why some options are both better and worse than others. Memory & Cognition, 21, 546-556.

Shafir, E. (1993). Choosing versus rejecting: Why some options are both better and worse than others. Memory & Cognition, 21, 546-556.
}
