% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Tversky.1}
\alias{varfun.Tversky.1}
\title{varfun.Tversky.1}
\usage{
varfun.Tversky.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#16_tversky}
}
\section{Variables}{

tver1.1=choice ($250 wall hanging condition, yes=1, no=2);
tver2.1=choice ($30 wall hanging cond, yes=1, no=2).

Materials and procedure. Participants will receive one of two scenarios from the original with dollar amounts approximately adjusted for inflation and the consumer items being replaced with a ceramic vase and a wall hanging.

tver1.1 Imagine that you are about to purchase a ceramic vase for $30, and a wall hanging for $250. The salesman informs you that the wall hanging you wish to buy is on sale for $240 at the other branch of the store, located 20 minutes drive away. Would you make the trip to the other store?
m  Yes, I would go to the other branch. (1)
m  No, I would not go to the other branch. (2)

tver2.1 Imagine that you are about to purchase a ceramic vase for $250, and a wall hanging for $30. The salesman informs you that the wall hanging you wish to buy is on sale for $20 at the other branch of the store, located 20 minutes drive away. Would you make the trip to the other store?
m  Yes, I would go to the other branch. (1)
m  No, I would not go to the other branch. (2)
}

\references{
Tversky, A., Kahneman, D. (1981). The framing of decisions and the psychology of choice. Science, 211, 453-458.
}
