% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.vanLange.1}
\alias{varfun.vanLange.1}
\alias{varfun.vanLange.1}
\alias{varfun.vanLange.1}
\title{van.Lange.1}
\usage{
varfun.vanLange.1(vars)

varfun.vanLange.1(vars)

varfun.vanLange.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis A \code{list} object containing fields \strong{SVO}, \strong{Siblings} and \strong{N}

Dataset ready for analysis A \code{list} object containing fields \strong{SVO}, \strong{Siblings} and \strong{N}

Dataset ready for analysis A \code{list} object containing fields \strong{SVO}, \strong{Siblings} and \strong{N}
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#10_van_lange}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#10_van_lange}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#10_van_lange}
}
\section{Variables}{

   van.p.1.2_1 TO van.p.1.2_6 are the items of SVO measure.

  murphy et al. (2011) scoring: SVO degress=arctan [(mean Alloc other - 50)/(mean Allocation self - 50)].

  See SVO codes (this doc) for the list of paired amounts
  van.p2.1_1_TEXT= # of older siblings;
  van.p2.1_2_TEXT= # of younger siblings.


   van.p.1.2_1 TO van.p.1.2_6 are the items of SVO measure.

  murphy et al. (2011) scoring: SVO degress=arctan [(mean Alloc other - 50)/(mean Allocation self - 50)].

  See SVO codes (this doc) for the list of paired amounts
  van.p2.1_1_TEXT= # of older siblings;
  van.p2.1_2_TEXT= # of younger siblings.


   van.p.1.2_1 TO van.p.1.2_6 are the items of SVO measure.

  murphy et al. (2011) scoring: SVO degress=arctan [(mean Alloc other - 50)/(mean Allocation self - 50)].

  See SVO codes (this doc) for the list of paired amounts
  van.p2.1_1_TEXT= # of older siblings;
  van.p2.1_2_TEXT= # of younger siblings.
}

