% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{varfun.Zhong.3}
\alias{varfun.Zhong.3}
\title{varfun.Zhong.3}
\usage{
varfun.Zhong.3(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#23_zhong}
}
\section{Variables}{

zhon1.1= unethical condition;
zhon2.1=ethical condition;

zhon.dv.1_1 TO zhon.dv.1_10=desirability of products (both conditions);
higher numbers=higher desirability;

Products list:

Clean:
Dove shower soap (zhon.dv.1_2),
Crest toothpaste (zhon.dv.1_3),
Windex glass cleaner (zhon.dv.1_7),
Lysol countertop disinfectant (zhon.dv.1_8),
Tide laundry detergent (zhon.dv.1_10)

Not-clean:
Post-it notes (zhon.dv.1_1),
Nantucket Nectars juice (zhon.dv.1_4),
Energizer batteries (zhon.dv.1_5),
Sony cd cases (zhon.dv.1_6),
Snickers candy bar (zhon.dv.1_9),
}

\references{
Zhong, C. B., & Liljenquist, K. (2006). Washing away your sins: Threatened morality and physical cleansing. Science, 313, 1451???1452.
}
