% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRedsRutils.R
\name{z.test}
\alias{z.test}
\title{z.test}
\usage{
z.test(x = 0, mu = 0, pi = NULL, N = 0, sigma = 1,
  proportion = FALSE, alternative = "two.sided")
}
\arguments{
\item{x}{Observed value}

\item{mu}{Population mean under H_{0}}

\item{pi}{Population proportion under H_{0}}

\item{N}{Sample size}

\item{sigma}{Population standard deviation}

\item{proportion}{Observed proportion}

\item{alternative}{}
}
\description{
An `htest` implentation of the z test for an abserved mean or proportion.
}
