% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{decide.analysis}
\alias{decide.analysis}
\title{decide.analysis}
\usage{
decide.analysis(ML2.key, studies = NA, tp = NA)
}
\arguments{
\item{ML2.key}{Key}

\item{studies}{studies}

\item{tp}{analysis type}
}
\value{
list of studies
}
\description{
decide.analysis
}
