% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/fRedsRutils.R
\name{disp}
\alias{disp}
\alias{disp}
\title{disp}
\usage{
disp(message = "Hello world!", header = "disp", footer = TRUE)

disp(message = "Hello world!", header = "disp", footer = TRUE)
}
\arguments{
\item{message}{A message to be displayed in the Console.}

\item{header}{Print a header of '~' symbols (=\code{TRUE}), or '~' symbols with few words of text (=\code{character vector})}

\item{footer}{Print a footer '~' symbols.}

\item{message}{A message to be displayed in the Console.}

\item{header}{Print a header of '~' symbols (=\code{TRUE}), or '~' symbols with few words of text (=\code{character vector})}

\item{footer}{Print a footer '~' symbols.}
}
\description{
Displays easy-to-spot text in the Console.

Displays easy-to-spot text in the Console.
}
\author{
Fred Hasselman

Fred Hasselman
}
