% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{generateOutput}
\alias{generateOutput}
\title{generateOutput}
\usage{
generateOutput(describe = describe, var.lor = NA,
  runningGroup = "None", runningAnalysis = "None", stat.params = NA)
}
\arguments{
\item{describe}{The output of function \link{\code{get.descriptives}}}

\item{var.lor}{Variance of log OR effect size}

\item{runningGroup}{The group on which the statistics in `describe` were calculated (can be: 'all', a source name, or a presentation order)}

\item{runningAnalysis}{The analysis that generated the statistics in `describe`}
}
\description{
generateOutput
}
\author{
Fred Hasselman
}
