% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R
\name{get.descriptives}
\alias{get.descriptives}
\title{get.descriptives}
\usage{
get.descriptives(stat.test, vars, keytable)
}
\arguments{
\item{stat.test}{Output of the analysis}

\item{vars}{Data used}

\item{keytable}{masteRkey}
}
\description{
get.descriptives
}
