% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manylabRs_SOURCE.R
\name{get.plotly}
\alias{get.plotly}
\title{get.plotly}
\usage{
get.plotly(data, analysis_url)
}
\arguments{
\item{data}{Dataframe with ML2 testresutls and ESCI output.}

\item{analysis_url}{url}
}
\description{
Get a plotly plot.
}
