% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R
\name{get.sourceData}
\alias{get.sourceData}
\alias{get.sourceData}
\title{get.sourceData}
\usage{
get.sourceData(ML2.id, ML2.df, ML2.in)

get.sourceData(ML2.id, ML2.df, ML2.in)
}
\arguments{
\item{ML2.id}{Internal}

\item{ML2.df}{Internal}

\item{ML2.in}{Inernal}

\item{ML2.id}{Internal}

\item{ML2.df}{Internal}

\item{ML2.in}{Inernal}
}
\value{
A list with fields \code{study.vars} (data organised according to the \code{masteRkey} spreadsheet), \code{study.vars/labels}, \code{N}, and \code{RawDataFilter}(raw data, unfiltered).

A list with fields \code{study.vars} (data organised according to the \code{masteRkey} spreadsheet), \code{study.vars/labels}, \code{N}, and \code{RawDataFilter}(raw data, unfiltered).
}
\description{
get.sourceData

get.sourceData
}
