% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/ML2_variable_functions.R
\name{varfun.Bauer.1}
\alias{varfun.Bauer.1}
\alias{varfun.Bauer.1}
\title{varfun.Bauer.1}
\usage{
varfun.Bauer.1(vars)

varfun.Bauer.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#6_bauer}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#6_bauer}
}
