% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/ML2_variable_functions.R
\name{varfun.Norenzayan.1}
\alias{varfun.Norenzayan.1}
\alias{varfun.Norenzayan.1}
\alias{varfun.Norenzayan.1}
\title{varfun.Norenzayan.1}
\usage{
varfun.Norenzayan.1(vars)

varfun.Norenzayan.1(vars)

varfun.Norenzayan.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#20_norenzayan}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#20_norenzayan}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#20_norenzayan}
}
\section{Variables}{

 nore1.1 TO nore1.20 provide choices ("belong to" condition)

 nore2.1 to nore2.20 provide choices ("similar to" condition)


 nore1.1 TO nore1.20 provide choices ("belong to" condition)

 nore2.1 to nore2.20 provide choices ("similar to" condition)


 nore1.1 TO nore1.20 provide choices ("belong to" condition)

 nore2.1 to nore2.20 provide choices ("similar to" condition)
}

\references{
Norenzayan, A., Smith, E. E., Kim, B. J., & Nisbett, R. E. (2002). Cultural preferences for formal versus intuitive reasoning. Cognitive Science, 26, 653-684.

Norenzayan, A., Smith, E. E., Kim, B. J., & Nisbett, R. E. (2002). Cultural preferences for formal versus intuitive reasoning. Cognitive Science, 26, 653-684.

Norenzayan, A., Smith, E. E., Kim, B. J., & Nisbett, R. E. (2002). Cultural preferences for formal versus intuitive reasoning. Cognitive Science, 26, 653-684.
}
