% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Ross.1}
\alias{varfun.Ross.1}
\title{varfun.Ross.1}
\usage{
varfun.Ross.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#13_ross1}
}
\section{Variables}{

ross.s1.1 = percentage of peers;
ross.s1.2 = you; values: 1=sign; 2=refuse
}

