% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML2_variable_functions.R
\name{varfun.Ross.2}
\alias{varfun.Ross.2}
\title{varfun.Ross.2}
\usage{
varfun.Ross.2(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#14_ross2}
}
\section{Variables}{

ross.s2.1=percentage of peers;
ross.s2.2=you; values: 1=Pay; 2=Appear in court
}

