% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII.R, R/ML2_variable_functions.R
\name{varfun.Zaval.1}
\alias{varfun.Zaval.1}
\alias{varfun.Zaval.1}
\title{varfun.Zaval.1}
\usage{
varfun.Zaval.1(vars)

varfun.Zaval.1(vars)
}
\arguments{
\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}

\item{vars}{A list object generated by \code{\link{get.sourceData}} containing cleaned data and variable labels.}
}
\value{
Dataset ready for analysis

Dataset ready for analysis
}
\description{
\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#26_zaval}

\url{https://manylabsopenscience.github.io/ML2_PoPS_proposal#26_zaval}
}
\section{Variables}{

zav1.1 TO zav1.13 provide COLD primes.
zav2.1 TO zav2.13 provide HEAT primes.
zav.dv.2=belief;
zav.dv.3=concern;
higher numbers=higher belief/concern.


zav1.1 TO zav1.13 provide COLD primes.
zav2.1 TO zav2.13 provide HEAT primes.
zav.dv.2=belief;
zav.dv.3=concern;
higher numbers=higher belief/concern.
}

\references{
Zaval, L., Keenan, E. A., Johnson, E. J., & Weber, E. U. (2014). How warm days increase belief in global warming. \strong{Nature Climate Change}, 4, 143-147.

Zaval, L., Keenan, E. A., Johnson, E. J., & Weber, E. U. (2014). How warm days increase belief in global warming. \strong{Nature Climate Change}, 4, 143-147.
}
