% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-3PR_ASCII 2.R, R/C-3PR_ASCII.R,
%   R/fRedsRutils.R
\name{vioQtile}
\alias{vioQtile}
\alias{vioQtile}
\alias{vioQtile}
\title{vioQtile}
\usage{
vioQtile(gg = NULL, qtiles = NULL, probs = seq(0, 1, 0.25),
  labels = paste(probs[-1] * 100), withData = FALSE)

vioQtile(gg = NULL, qtiles = NULL, probs = seq(0, 1, 0.25),
  labels = paste(probs[-1] * 100), withData = FALSE)

vioQtile(gg = NULL, qtiles = NULL, probs = seq(0, 1, 0.25),
  labels = paste(probs[-1] * 100), withData = FALSE)
}
\arguments{
\item{gg}{A ggplot.}

\item{qtiles}{Quantiles.}

\item{probs}{Probabilities.}

\item{labels}{Labels.}

\item{withData}{Return Data.}

\item{gg}{A ggplot.}

\item{qtiles}{Quantiles.}

\item{probs}{Probabilities.}

\item{labels}{Labels.}

\item{withData}{Return Data.}

\item{gg}{A ggplot.}

\item{qtiles}{Quantiles.}

\item{probs}{Probabilities.}

\item{labels}{Labels.}

\item{withData}{Return Data.}
}
\description{
vioQtile

vioQtile

vioQtile
}
\details{
This is adapted from: \url{http://stackoverflow.com/questions/22278951/combining-violin-plot-with-box-plot}

 \strong{Changed:}
 Deal with 'empty' quantile groups
 Deal with original data
 More input, more output

This is adapted from: \url{http://stackoverflow.com/questions/22278951/combining-violin-plot-with-box-plot}

 \strong{Changed:}
 Deal with 'empty' quantile groups
 Deal with original data
 More input, more output

This is adapted from: \url{http://stackoverflow.com/questions/22278951/combining-violin-plot-with-box-plot}

 \strong{Changed:}
 Deal with 'empty' quantile groups
 Deal with original data
 More input, more output
}
